+++ src/main/kotlin/com/gitee/authentication/GECredentials.kt
@@ -12,6 +12,9 @@ class GECredentials(
 ) : CredentialsWithRefresh {
 
   override fun isAccessTokenValid(): Boolean {
+    // 私有令牌长期有效
+    if (tokenType == "private" && accessToken != "") return true
+
     return System.currentTimeMillis() < (createdAt + expiresIn) * 1000
   }
 
@@ -19,7 +22,11 @@ class GECredentials(
     private const val empty_str = ""
     val EmptyCredentials = GECredentials(empty_str, empty_str, 0, empty_str, empty_str, 0)
 
-    fun createCredentials(accessToken: String, refreshToken: String) =
-      GECredentials(accessToken, refreshToken, 86400, "bearer", GEAccountsUtil.APP_CLIENT_SCOPE, System.currentTimeMillis() / 1000)
+    fun createCredentials(accessToken: String, refreshToken: String, isPrivateToken: Boolean = false) =
+      if (isPrivateToken)
+        GECredentials(accessToken, "", 0, "private", GEAccountsUtil.APP_CLIENT_SCOPE, System.currentTimeMillis() / 1000)
+      else
+        GECredentials(accessToken, refreshToken, 86400, "bearer", GEAccountsUtil.APP_CLIENT_SCOPE, System.currentTimeMillis() / 1000)
+
   }
 }